- Reworked ImagePyramids  * Update and the pyramid data structure are now two separate entities  * Added strong typing to UpdatePyramid  * Updated everything which referenced a pyramid  * Created PyramidOps for derivative and filter
- ScaleSpacePyramid
 * New equation for approximating the scale space
 * Simplified interface

- Evaluation
 * Pushed KLT track video into a single file

- Misc
 * Various refactoring
 * Fixed some unit tests

+++ ip/src/gecv/core/image/GeneralizedImageOps.java
@@ -239,9 +239,9 @@ public class GeneralizedImageOps {
 		} else if ( ImageSInt64.class.isAssignableFrom(img.getClass()) ) {
 			ImageTestingOps.fill((ImageSInt64)img,(int)value);
 		} else if (ImageFloat32.class.isAssignableFrom(img.getClass()) ) {
-			ImageTestingOps.fill((ImageFloat32)img,(int)value);
+			ImageTestingOps.fill((ImageFloat32)img,(float)value);
 		} else if (ImageFloat64.class.isAssignableFrom(img.getClass()) ) {
-			ImageTestingOps.fill((ImageFloat64)img,(int)value);
+			ImageTestingOps.fill((ImageFloat64)img,value);
 		} else {
 			throw new IllegalArgumentException("Unknown or incompatible image type: " + img.getClass().getSimpleName());
 		}
